+++ AndroidClient/src/main/java/com/tom/meeter/context/profile/repository/user/database/UserDatabase.java
@@ -1,9 +1,9 @@
-package com.tom.meeter.context.profile.user.database;
+package com.tom.meeter.context.profile.repository.user.database;
 
 import androidx.room.Database;
 import androidx.room.RoomDatabase;
 
-import com.tom.meeter.context.profile.user.domain.User;
+import com.tom.meeter.context.profile.repository.user.domain.User;
 
 @Database(entities = {User.class}, version = 2)
 public abstract class UserDatabase extends RoomDatabase {
